refactor(data-value-mutations): move inputs to new mutation hooks
