+++ src/data-workspace/inputs/long-text.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import { useField } from 'react-final-form'
-import { useDataValueMutation } from '../data-entry-cell/use-data-value-mutation.js'
 import styles from './inputs.module.css'
+import useDataValueMutation from './use-data-value-mutation.js'
 import { InputPropTypes } from './utils.js'
 
 export const LongText = ({ fieldname, dataValueParams, setSyncStatus }) => {
