+++ src/data-workspace/inputs/true-only-checkbox.js
@@ -1,8 +1,8 @@
 import { Checkbox } from '@dhis2/ui'
 import React, { useState } from 'react'
 import { useField } from 'react-final-form'
-import { useDataValueMutation } from '../data-entry-cell/use-data-value-mutation.js'
 import styles from './inputs.module.css'
+import useDataValueMutation from './use-data-value-mutation.js'
 import { convertCallbackSignatures, InputPropTypes } from './utils.js'
 
 export const TrueOnlyCheckbox = ({
